+++ aldryn_newsblog/tests/test_cms_wizards.py
@@ -1,5 +1,6 @@
 from aldryn_newsblog.cms_wizards import CreateNewsBlogArticleForm
-from aldryn_newsblog.tests import NewsBlogTestCase
+
+from .mixins import NewsBlogTestCase
 
 
 class CreateNewsBlogArticleFormTestCase(NewsBlogTestCase):
@@ -19,8 +20,8 @@ class CreateNewsBlogArticleFormTestCase(NewsBlogTestCase):
 
         article = form.save()
         self.assertTrue(article.__class__.objects.filter(id=article.id).exists())
-        self.assertEqual(article.content.get_plugins('en').count(), 1)
-        plugin = article.content.get_plugins('en').get()
+        self.assertEqual(article.content_placeholder.get_plugins('en').count(), 1)
+        plugin = article.content_placeholder.get_plugins('en').get()
         self.assertEqual(plugin.plugin_type, 'TextPlugin')
         self.assertEqual(plugin.djangocms_text_text.body, 'My super content')
 
@@ -29,18 +30,18 @@ class CreateNewsBlogArticleFormTestCase(NewsBlogTestCase):
 
         article = form.save()
         self.assertTrue(article.__class__.objects.filter(id=article.id).exists())
-        self.assertFalse(article.content.get_plugins('en').exists())
+        self.assertFalse(article.content_placeholder.get_plugins('en').exists())
 
     def test_article_is_saved_with_content_without_plugin_permission(self):
         form = self.get_form(has_content=True, has_permission=False)
 
         article = form.save()
         self.assertTrue(article.__class__.objects.filter(id=article.id).exists())
-        self.assertFalse(article.content.get_plugins('en').exists())
+        self.assertFalse(article.content_placeholder.get_plugins('en').exists())
 
     def test_article_is_saved_without_content_without_plugin_permission(self):
         form = self.get_form(has_content=False, has_permission=False)
 
         article = form.save()
         self.assertTrue(article.__class__.objects.filter(id=article.id).exists())
-        self.assertFalse(article.content.get_plugins('en').exists())
+        self.assertFalse(article.content_placeholder.get_plugins('en').exists())