+++ aldryn_newsblog/tests/test_managers.py
@@ -1,6 +1,6 @@
 from aldryn_newsblog.models import Article
 
-from . import NewsBlogTestCase
+from .mixins import NewsBlogTestCase
 
 
 class TestManagers(NewsBlogTestCase):